Add new shroud part, re-do gear,  assm of gear+shroud
Added a new shroud part to hold the beads in. Slightly modified the gear teeth length to make it slightly longer (and the radius slightly to make an integer number of instances of teeth). The reason being is that the teeth on the previous gear version were slightly too close together and din't mesh completely. Also slightly increased the radius of the teeth hole in order to better fit the beads.